// src/generated/model/batchUpsertRecordsResponseError.ts
/**
 * Search.io API
 * Search.io offers a search and discovery service with Neuralsearch®, the world\'s first instant AI search technology. Businesses of all sizes use Search.io to build site search and discovery solutions that maximize e-commerce revenue, optimize on-site customer experience, and scale their online presence.
 *
 * The version of the OpenAPI document: v4
 * Contact: support@search.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";
import { Status } from "./status";

export class BatchUpsertRecordsResponseError {
  /**
   * Index of the record in `records` that this error corresponds to.
   */
  "index"?: number;
  "status"?: Status;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "index",
      baseName: "index",
      type: "number",
    },
    {
      name: "status",
      baseName: "status",
      type: "Status",
    },
  ];

  static getAttributeTypeMap() {
    return BatchUpsertRecordsResponseError.attributeTypeMap;
  }
}