// src/generated/model/getDefaultPipelineResponse.ts
/**
 * Sajari API
 * Sajari is a smart, highly-configurable, real-time search service that enables thousands of businesses worldwide to provide amazing search experiences on their websites, stores, and applications.
 *
 * The version of the OpenAPI document: v4
 * Contact: support@sajari.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";

export class GetDefaultPipelineResponse {
  /**
   * The name of the pipeline to use when not otherwise specified.
   */
  "pipeline"?: string;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "pipeline",
      baseName: "pipeline",
      type: "string",
    },
  ];

  static getAttributeTypeMap() {
    return GetDefaultPipelineResponse.attributeTypeMap;
  }
}