// src/generated/model/listCollectionsResponse.ts
/**
 * Search.io API
 * Search.io offers a search and discovery service with Neuralsearch®, the world\'s first instant AI search technology. Businesses of all sizes use Search.io to build site search and discovery solutions that maximize e-commerce revenue, optimize on-site customer experience, and scale their online presence.
 *
 * The version of the OpenAPI document: v4
 * Contact: support@search.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from "./models";
import { Collection } from "./collection";

export class ListCollectionsResponse {
  /**
   * The collections from the specified account.
   */
  "collections"?: Array<Collection>;
  /**
   * A token, which can be sent as `page_token` to retrieve the next page.  If this field is omitted, there are no subsequent pages.
   */
  "nextPageToken"?: string;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: "collections",
      baseName: "collections",
      type: "Array<Collection>",
    },
    {
      name: "nextPageToken",
      baseName: "next_page_token",
      type: "string",
    },
  ];

  static getAttributeTypeMap() {
    return ListCollectionsResponse.attributeTypeMap;
  }
}